print fee invoice methods
I manually created and added students into the student arraylist at the start of the main method for testing purposes. We have to remove it before submitting the file.

I added a query for user input in the student menu; a helper method printFeeInvoice(String stuID) in the main class; and printInvoice() methods in the undergraduate, Ms, and PhD student classes.